Apr 2, 2025, 7:47 AM

+++ CT06_11/lesson11.py
@@ -9,7 +9,7 @@
 #     px <= 500: "Where are you getting this money from?!"
 #     px > 500: "Don't even think about it!"
 # RECAP
-px = int(input("What is the price of the item you want to buy?"))
+px = float(input("What is the price of the item you want to buy?"))
 
 if px <= 5:
     print("Sounds good!")
